Move file traversal methods into separate file
